+++ changelog.md
@@ -5,6 +5,7 @@
 ## Changes in release 6.4.1
 + Fixed payment fee calculation on checkout page
 + Re-added payment method title translations
++ Removed unused Sentry error tracking code
 + Fixed errors in logs when changing order statuses of Payments API orders
 
 ## Changes in release 6.4.0